backend/requirements.txt
google-genai==1.12.1
python-dotenv
supabase
letta
flask
flask-cors